Address review suggestions: Move Op classes to pennylane extenstions
